backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pydantic-settings==2.3.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
